requirements.txt
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
h11==0.14.0
hijri-converter==2.3.1
hijridate==2.3.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.0.1
pytz==2024.1
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
wcwidth==0.2.13
Flask==2.3.2
gunicorn==21.0.0
